+++ internal/pearserver/simplespace_add_member.go
@@ -1,0 +1,49 @@
+package pearserver
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/http"
+
+	"github.com/habitat-network/habitat/api/habitat"
+	"github.com/habitat-network/habitat/internal/authn"
+	"github.com/habitat-network/habitat/internal/httpx"
+	"github.com/habitat-network/habitat/internal/spaces"
+	habitat_syntax "github.com/habitat-network/habitat/internal/syntax"
+)
+
+// AddMember implements network.habitat.simplespace.addMember.
+func (p *PearServer) AddMember(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	var input habitat.NetworkHabitatSimplespaceAddMemberInput
+	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
+		httpx.WriteInvalidRequest(ctx, w, "failed to parse params", err)
+		return
+	}
+	spaceURI, ok := httpx.ParseSpaceURIInput(ctx, w, input.Space, "space uri")
+	if !ok {
+		return
+	}
+	_, ok = p.validator.Request(
+		authn.WithMethods(authn.ValidatorMethodOAuth, authn.ValidatorMethodServiceAuth),
+		authn.WithSpace(spaceURI, habitat_syntax.SpaceRoleManager),
+	).Validate(w, r)
+	if !ok {
+		return
+	}
+	memberDID, ok := httpx.ParseDIDInput(ctx, w, input.Did, "did")
+	if !ok {
+		return
+	}
+	err := p.simpleStore.AddMember(r.Context(), spaceURI, memberDID)
+	if errors.Is(err, spaces.ErrSpaceNotFound) {
+		httpx.WriteSpaceNotFound(ctx, w, err)
+		return
+	} else if errors.Is(err, spaces.ErrUserAlreadyMember) {
+		return
+	} else if err != nil {
+		httpx.WriteServerError(ctx, w, fmt.Errorf("add member: %w", err))
+		return
+	}
+}
