+++ internal/pearserver/space_test_constants_test.go
@@ -8,5 +8,8 @@ var (
 	org     = syntax.DID("did:plc:org")
 	owner   = syntax.DID("did:plc:owner")
 	alice   = syntax.DID("did:plc:alice")
+	bob     = syntax.DID("did:plc:bob")
+	admin   = syntax.DID("did:plc:admin")
 	groupTp = syntax.NSID("network.habitat.group")
+	docsTp  = syntax.NSID("network.habitat.docs")
 )
